Modify custom bool typedef
I'm not sure why I had it as a uint8 instead of a int32.

+++ common/src/repetition_tester.h
@@ -2,7 +2,7 @@
 #define REPETITION_TESTER_H
 
 #include <stdint.h>
-typedef int8_t rt__b32;
+typedef int32_t rt__b32;
 
 #define RT__TRUE 1
 #define RT__FALSE 0
